Move gql directory
