+++ src/components/class_manager/ClassCreator.js
@@ -6,8 +6,8 @@ import CircularProgress from '@material-ui/core/CircularProgress';
 import IconButton from '@material-ui/core/IconButton';
 import AddCircleIcon from '@material-ui/icons/AddCircle';
 import { useMutation } from '@apollo/react-hooks';
-import { GET_PERIODS } from '../gql/queries/Period';
-import { CREATE_PERIOD } from '../gql/mutations/Period';
+import { GET_PERIODS } from '../../gql/queries/Period';
+import { CREATE_PERIOD } from '../../gql/mutations/Period';
 import { NotificationsContext } from '../app/notifications/NotificationsContext';
 import parseError from '../../util/parseError';
