+++ src/components/class_manager/ClassEdit.js
@@ -6,8 +6,8 @@ import CloseIcon from '@material-ui/icons/Close';
 import CheckIcon from '@material-ui/icons/Check';
 import { useQuery, useMutation } from '@apollo/react-hooks';
 import GlobalLoader from '../app/GlobalLoader';
-import { GET_PERIOD } from '../gql/queries/Period';
-import { EDIT_PERIOD, DELETE_PERIOD } from '../gql/mutations/Period';
+import { GET_PERIOD } from '../../gql/queries/Period';
+import { EDIT_PERIOD, DELETE_PERIOD } from '../../gql/mutations/Period';
 import EditIcon from '@material-ui/icons/Edit';
 import DeleteIcon from '@material-ui/icons/Delete';
 import ErrorModal from '.././app/ErrorModal';
