+++ src/components/class_manager/assignments/AssignmentResults.js
@@ -8,7 +8,7 @@ import StudentAssignmentShow from './student_assignment_show';
 import {
   GET_ASSIGNMENT_RESULTS,
   GET_ASSIGNMENT,
-} from '../../gql/queries/Assignment';
+} from '../../../gql/queries/Assignment';
 import MaterialTable from '../../table/MaterialTable';
 import GlobalLoader from '../../app/GlobalLoader';
