+++ src/components/decks/deck_editor/CurrentDeck.js
@@ -16,8 +16,8 @@ import { makeStyles } from '@material-ui/core/styles';
 import Confirmation from './Confirmation';
 import { CurrentDeckContext } from '../CurrentDeckContext';
 import CardsContainer from './CardsContainer';
-import { GET_DECK } from '../../gql/queries/Deck';
-import { DELETE_DECK } from '../../gql/mutations/Deck';
+import { GET_DECK } from '../../../gql/queries/Deck';
+import { DELETE_DECK } from '../../../gql/mutations/Deck';
 import { NotificationsContext } from '../../app/notifications/NotificationsContext';
 
 const useStyles = makeStyles(theme => ({
