+++ src/components/questions/question_form/EditForm.js
@@ -13,7 +13,7 @@ import Button from '@material-ui/core/Button';
 import Form from './Form';
 import { QuestionFormProvider } from './QuestionFormContext';
 import { CurrentDeckContext } from '../../decks/CurrentDeckContext';
-import { UPDATE_QUESTION } from '../../gql/mutations/Question';
+import { UPDATE_QUESTION } from '../../../gql/mutations/Question';
 import GlobalLoader from '../../app/GlobalLoader';
 
 const UpdateForm = ({ open, setOpen, card }) => {
