+++ src/components/standards/StandardsManager.js
@@ -4,9 +4,9 @@ import { useQuery, useMutation } from '@apollo/react-hooks';
 import TextField from '@material-ui/core/TextField';
 import MaterialTable from '../table/MaterialTable';
 import { NotificationsContext } from '../app/notifications/NotificationsContext';
-import { GET_STANDARDS_WITH_CATEGORIES } from '../gql/queries/Standard';
-import { GET_STANDARDS_CATEGORIES } from '../gql/queries/StandardsCategory';
-import { DELETE_STANDARD, EDIT_STANDARD } from '../gql/mutations/Standard';
+import { GET_STANDARDS_WITH_CATEGORIES } from '../../gql/queries/Standard';
+import { GET_STANDARDS_CATEGORIES } from '../../gql/queries/StandardsCategory';
+import { DELETE_STANDARD, EDIT_STANDARD } from '../../gql/mutations/Standard';
 import GlobalLoader from '../app/GlobalLoader';
 import CategoriesManager from './CategoriesManager';
 import StandardsForm from './StandardsForm';
