+++ src/gql/queries/Period.js
