+++ src/gql/queries/Standard.js
